getIsOnPaidPlan
